requirements.txt
python-dotenv
openai>=1.0.0
pandas
tqdm
scikit-learn
